Separate the components and containers from each other
